# apps/backend/doc/openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce API
  description: |
    A comprehensive e-commerce API built with Express.js, TypeScript, and Prisma.
    This API provides authentication, product management, and other e-commerce functionalities.

    ## Features
    - JWT-based authentication
    - Google OAuth integration
    - Product catalog management
    - Admin role-based access control
    - Comprehensive error handling

    ## Authentication
    Most endpoints require authentication via JWT token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```
  version: 1.0.0
  contact:
    name: E-commerce API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/wishlist:
    get:
      tags:
        - Wishlist
      summary: Get user wishlist
      description: Retrieve the current user's wishlist
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Wishlist retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Wishlist retrieved successfully'
                  data:
                    $ref: '#/components/schemas/Wishlist'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Wishlist
      summary: Add item to wishlist
      description: Add a product to the current user's wishlist
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
              properties:
                productId:
                  type: string
                  format: uuid
                  example: '123e4567-e89b-12d3-a456-426614174000'
      responses:
        '201':
          description: Item added to wishlist successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Item added to wishlist successfully'
                  data:
                    $ref: '#/components/schemas/WishlistItem'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Product already in wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Wishlist
      summary: Clear wishlist
      description: Remove all items from the current user's wishlist
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Wishlist cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Wishlist cleared successfully'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/wishlist/{productId}:
    delete:
      tags:
        - Wishlist
      summary: Remove item from wishlist
      description: Remove a specific product from the current user's wishlist
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product ID
      responses:
        '200':
          description: Item removed from wishlist successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Item removed from wishlist successfully'
        '400':
          description: Invalid product ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Product not found in wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/wishlist/{productId}/move-to-cart:
    post:
      tags:
        - Wishlist
      summary: Move item to cart
      description: Move a product from wishlist to shopping cart
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
                  example: '123e4567-e89b-12d3-a456-426614174000'
                quantity:
                  type: integer
                  minimum: 1
                  example: 1
      responses:
        '200':
          description: Item moved to cart successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Item moved to cart successfully'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Product not found in wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/wishlist/{productId}/status:
    get:
      tags:
        - Wishlist
      summary: Check wishlist status
      description: Check if a product is in the current user's wishlist
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product ID
      responses:
        '200':
          description: Wishlist status checked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Wishlist status checked successfully'
                  data:
                    type: object
                    properties:
                      productId:
                        type: string
                        format: uuid
                      isInWishlist:
                        type: boolean
        '400':
          description: Invalid product ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/wishlist/count:
    get:
      tags:
        - Wishlist
      summary: Get wishlist count
      description: Get the number of items in the current user's wishlist
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Wishlist count retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Wishlist count retrieved successfully'
                  data:
                    type: object
                    properties:
                      count:
                        type: integer
                        example: 5
        '401':
          $ref: '#/components/responses/Unauthorized'

  # User Preferences Endpoints
  /api/preferences:
    get:
      tags:
        - Preferences
      summary: Get user preferences
      description: Retrieve the current user's preferences
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Preferences retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Preferences retrieved successfully'
                  data:
                    $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Preferences
      summary: Update user preferences
      description: Update the current user's preferences
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Preferences updated successfully'
                  data:
                    $ref: '#/components/schemas/UserPreferences'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/preferences/notifications:
    get:
      tags:
        - Preferences
      summary: Get notification settings
      description: Retrieve the current user's notification settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notification settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Notification settings retrieved successfully'
                  data:
                    $ref: '#/components/schemas/NotificationSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Preferences
      summary: Update notification settings
      description: Update the current user's notification settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettings'
      responses:
        '200':
          description: Notification settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Notification settings updated successfully'
                  data:
                    $ref: '#/components/schemas/UserPreferences'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/preferences/display:
    get:
      tags:
        - Preferences
      summary: Get display settings
      description: Retrieve the current user's display settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Display settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Display settings retrieved successfully'
                  data:
                    $ref: '#/components/schemas/DisplaySettings'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Preferences
      summary: Update display settings
      description: Update the current user's display settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisplaySettings'
      responses:
        '200':
          description: Display settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Display settings updated successfully'
                  data:
                    $ref: '#/components/schemas/UserPreferences'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/preferences/reset:
    post:
      tags:
        - Preferences
      summary: Reset preferences to default
      description: Reset the current user's preferences to default values
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Preferences reset to default successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Preferences reset to default successfully'
                  data:
                    $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: 'Bad Request'
        message:
          type: string
          example: 'Invalid input provided'
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: 'Validation failed'
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              expected:
                type: string
              received:
                type: string
              path:
                type: array
                items:
                  type: string
              message:
                type: string

    # User Profile Schemas
...
    Wishlist:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WishlistItem'
        totalCount:
          type: integer
          example: 5
        totalValue:
          type: number
          format: float
          example: 299.99
    WishlistItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        createdAt:
          type: string
          format: date-time
        product:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: '123e4567-e89b-12d3-a456-426614174000'
            name:
              type: string
              example: 'iPhone 15'
            description:
              type: string
              example: 'Latest iPhone model'
            price:
              type: number
              format: float
              example: 999.99
            category:
              type: string
              example: 'Electronics'
            imageUrl:
              type: string
              nullable: true
              example: 'https://example.com/iphone15.jpg'
            stock:
              type: integer
              example: 10

    # User Preferences Schemas
    UserPreferences:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        emailNotifications:
          type: boolean
          example: true
        marketingEmails:
          type: boolean
          example: true
        orderNotifications:
          type: boolean
          example: true
        language:
          type: string
          example: 'en'
        currency:
          type: string
          example: 'USD'
        timezone:
          type: string
          example: 'UTC'
        theme:
          type: string
          enum: [light, dark, auto]
          example: 'light'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdatePreferencesRequest:
      type: object
      properties:
        emailNotifications:
          type: boolean
          example: true
        marketingEmails:
          type: boolean
          example: true
        orderNotifications:
          type: boolean
          example: true
        language:
          type: string
          example: 'en'
        currency:
          type: string
          example: 'USD'
        timezone:
          type: string
          example: 'UTC'
        theme:
          type: string
          enum: [light, dark, auto]
          example: 'light'
    NotificationSettings:
      type: object
      required:
        - emailNotifications
        - marketingEmails
        - orderNotifications
      properties:
        emailNotifications:
          type: boolean
          example: true
        marketingEmails:
          type: boolean
          example: true
        orderNotifications:
          type: boolean
          example: true
    DisplaySettings:
      type: object
      required:
        - language
        - currency
        - timezone
        - theme
      properties:
        language:
          type: string
          example: 'en'
        currency:
          type: string
          example: 'USD'
        timezone:
          type: string
          example: 'UTC'
        theme:
          type: string
          enum: [light, dark, auto]
          example: 'light'

    # Upload Schemas
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Unauthorized'
              message:
                type: string
                example: 'Invalid or missing token'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...